+++ mohaemukzip/Secret.xcconfig
@@ -8,4 +8,4 @@
 
 
 BASE_URL = https:/$()/api.mohaemukzip.site
-ACCESS_TOKEN =eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwiYXV0aCI6IlJPTEVfQURNSU4iLCJpYXQiOjE3NzAxMDIwNjAsImV4cCI6MTc3MDEwMzg2MH0.7dQ6p-66dByOhv3ebuODpt2d3-6qkY7c7Lf4CLgbF_g
+ACCESS_TOKEN =eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwiYXV0aCI6IlJPTEVfQURNSU4iLCJpYXQiOjE3NzAxMjkxMTAsImV4cCI6MTc3MDEzMDkxMH0.zMHKUOwRdml3AjnUDCRT8hdBHlw7mWq5uSKz5ednNXw
