+++ mohaemukzip/Sources/Feature/Profile/Views/ProfileView.swift
@@ -435,335 +435,6 @@ private struct ProfileSimpleRow: View {
 
 
 
-// MARK: - 최근 조회한 레시피 목록 화면
-
-struct RecentlyViewedRecipesView: View {
-
-    // MARK: - Properties
-
-    @StateObject private var viewModel = RecentlyViewedRecipesViewModel()
-
-    // MARK: - Body
-
-    var body: some View {
-        VStack(spacing: 0) {
-            listHeader(title: "최근 조회한 레시피")
-
-            if viewModel.isLoading {
-                ProgressView().padding(.top, 20)
-            } else if viewModel.items.isEmpty {
-                Text("최근 조회한 레시피가 없습니다.")
-                    .font(.system(size: 14, weight: .medium))
-                    .foregroundStyle(Color.gray)
-                    .frame(maxWidth: .infinity, maxHeight: .infinity)
-            } else {
-                List {
-                    ForEach(viewModel.items) { item in
-                        RecipeListRow(recipe: item)
-                    }
-                }
-                .listStyle(.plain)
-            }
-        }
-        .navigationBarHidden(true)
-        .task {
-            // ✅ 최근 조회 목록 최초 로드
-            // - 콘솔 로그: [RecentlyViewedRecipesVM] fetch START/SUCCESS/FAIL
-            viewModel.fetch()
-        }
-        .alert(
-            "오류",
-            isPresented: Binding(
-                get: { viewModel.errorMessage != nil },
-                set: { if !$0 { viewModel.errorMessage = nil } }
-            )
-        ) {
-            Button("확인", role: .cancel) { }
-        } message: {
-            Text(viewModel.errorMessage ?? "")
-        }
-    }
-
-    private func listHeader(title: String) -> some View {
-        HStack {
-            Image(systemName: "chevron.left")
-                .font(.system(size: 16, weight: .semibold))
-                .foregroundStyle(Color.black)
-                .padding(.leading, 16)
-                .padding(.vertical, 12)
-                .onTapGesture {
-                    // NavigationStack에서 pop
-                    UIApplication.shared.sendAction(
-                        #selector(UINavigationController.popViewController(animated:)),
-                        to: nil,
-                        from: nil,
-                        for: nil
-                    )
-                }
-
-            Text(title)
-                .font(.system(size: 18, weight: .bold))
-
-            Spacer()
-        }
-    }
-}
-
-final class RecentlyViewedRecipesViewModel: ObservableObject {
-
-    // MARK: - Published
-
-    @Published private(set) var items: [ProfileRecipeCard] = []
-    @Published private(set) var isLoading: Bool = false
-    @Published var errorMessage: String? = nil
-
-    // MARK: - Functions
-
-    func fetch() {
-        guard !isLoading else { return }
-
-        isLoading = true
-        errorMessage = nil
-
-        print("[RecentlyViewedRecipesVM] ✅ fetch START")
-
-        ProfileService.shared.getRecentlyViewedRecipes { [weak self] result in
-            guard let self else { return }
-
-            switch result {
-            case let .success(dtoList):
-                let mapped = dtoList.map { ProfileRecipeCard.from(dto: $0) }
-
-                DispatchQueue.main.async {
-                    self.items = mapped
-                    self.isLoading = false
-                }
-
-                print("[RecentlyViewedRecipesVM] ✅ fetch SUCCESS | count=\(mapped.count)")
-
-            case let .failure(error):
-                DispatchQueue.main.async {
-                    self.isLoading = false
-                    self.errorMessage = error.localizedDescription
-                }
-
-                print("[RecentlyViewedRecipesVM] ❌ fetch FAIL | error=\(error)")
-            }
-        }
-    }
-}
-
-// MARK: - 저장한 레시피 목록 화면
-
-struct BookmarkedRecipesView: View {
-
-    // MARK: - Properties
-
-    @StateObject private var viewModel = BookmarkedRecipesViewModel()
-
-    // MARK: - Body
-
-    var body: some View {
-        VStack(spacing: 0) {
-            listHeader(title: "저장한 레시피")
-
-            if viewModel.isLoading && viewModel.items.isEmpty {
-                ProgressView().padding(.top, 20)
-            } else if viewModel.items.isEmpty {
-                Text("저장한 레시피가 없습니다.")
-                    .font(.system(size: 14, weight: .medium))
-                    .foregroundStyle(Color.gray)
-                    .frame(maxWidth: .infinity, maxHeight: .infinity)
-            } else {
-                List {
-                    ForEach(viewModel.items) { item in
-                        RecipeListRow(recipe: item)
-                            .onAppear {
-                                // 마지막 셀에 도달하면 다음 페이지 로드
-                                viewModel.loadNextIfNeeded(currentItem: item)
-                            }
-                    }
-
-                    if viewModel.isLoading && !viewModel.items.isEmpty {
-                        HStack {
-                            Spacer()
-                            ProgressView()
-                            Spacer()
-                        }
-                        .listRowSeparator(.hidden)
-                    }
-                }
-                .listStyle(.plain)
-            }
-        }
-        .navigationBarHidden(true)
-        .task {
-            // ✅ 저장한 레시피 목록 1페이지부터 로드
-            // - 페이지네이션은 마지막 셀이 보일 때(loadNextIfNeeded) 다음 페이지를 호출합니다.
-            // - 콘솔 로그: [BookmarkedRecipesVM] fetchPage START/SUCCESS/FAIL
-            viewModel.fetchFirstPage()
-        }
-        .alert(
-            "오류",
-            isPresented: Binding(
-                get: { viewModel.errorMessage != nil },
-                set: { if !$0 { viewModel.errorMessage = nil } }
-            )
-        ) {
-            Button("확인", role: .cancel) { }
-        } message: {
-            Text(viewModel.errorMessage ?? "")
-        }
-    }
-
-    private func listHeader(title: String) -> some View {
-        HStack {
-            Image(systemName: "chevron.left")
-                .font(.system(size: 16, weight: .semibold))
-                .foregroundStyle(Color.black)
-                .padding(.leading, 16)
-                .padding(.vertical, 12)
-                .onTapGesture {
-                    UIApplication.shared.sendAction(
-                        #selector(UINavigationController.popViewController(animated:)),
-                        to: nil,
-                        from: nil,
-                        for: nil
-                    )
-                }
-
-            Text(title)
-                .font(.system(size: 18, weight: .bold))
-
-            Spacer()
-        }
-    }
-}
-
-final class BookmarkedRecipesViewModel: ObservableObject {
-
-    // MARK: - Published
-
-    @Published private(set) var items: [ProfileRecipeCard] = []
-    @Published private(set) var isLoading: Bool = false
-    @Published var errorMessage: String? = nil
-
-    // MARK: - Pagination
-
-    private var currentPage: Int = 0
-    private var isLast: Bool = false
-
-    // MARK: - Functions
-
-    func fetchFirstPage() {
-        currentPage = 0
-        isLast = false
-        items = []
-        fetchPage(page: currentPage)
-    }
-
-    func loadNextIfNeeded(currentItem: ProfileRecipeCard) {
-        guard !isLoading, !isLast else { return }
-        guard let last = items.last else { return }
-
-        // 마지막 아이템이 화면에 나타났다면 다음 페이지 요청
-        if last.id == currentItem.id {
-            fetchPage(page: currentPage + 1)
-        }
-    }
-
-    private func fetchPage(page: Int) {
-        guard !isLoading else { return }
-
-        isLoading = true
-        errorMessage = nil
-
-        print("[BookmarkedRecipesVM] ✅ fetchPage START | page=\(page)")
-
-        ProfileService.shared.getBookmarkedRecipes(page: page) { [weak self] result in
-            guard let self else { return }
-
-            switch result {
-            case let .success(dtoPage):
-                let mapped = dtoPage.recipeList.map { ProfileRecipeCard.from(dto: $0) }
-
-                DispatchQueue.main.async {
-                    if page == 0 {
-                        self.items = mapped
-                    } else {
-                        self.items.append(contentsOf: mapped)
-                    }
-
-                    self.currentPage = page
-                    self.isLast = dtoPage.isLast
-                    self.isLoading = false
-                }
-
-                print("[BookmarkedRecipesVM] ✅ fetchPage SUCCESS | page=\(page), added=\(mapped.count), total=\(self.items.count), isLast=\(dtoPage.isLast)")
-
-            case let .failure(error):
-                DispatchQueue.main.async {
-                    self.isLoading = false
-                    self.errorMessage = error.localizedDescription
-                }
-
-                print("[BookmarkedRecipesVM] ❌ fetchPage FAIL | page=\(page), error=\(error)")
-            }
-        }
-    }
-}
-
-// MARK: - 리스트 Row 공용
-
-private struct RecipeListRow: View {
-
-    // MARK: - Properties
-
-    let recipe: ProfileRecipeCard
-
-    // MARK: - Body
-
-    var body: some View {
-        HStack(spacing: 12) {
-            ZStack(alignment: .bottomTrailing) {
-                ProfileYouTubeThumbnailView(videoId: recipe.videoId)
-                    .frame(width: 160, height: 96)
-                    .clipShape(RoundedRectangle(cornerRadius: 8))
-
-                Text(recipe.videoDurationText)
-                    .font(.system(size: 11, weight: .semibold))
-                    .foregroundStyle(Color.white)
-                    .padding(.horizontal, 6)
-                    .padding(.vertical, 3)
-                    .background(Color.black.opacity(0.75))
-                    .clipShape(RoundedRectangle(cornerRadius: 6))
-                    .padding(6)
-            }
-
-            VStack(alignment: .leading, spacing: 6) {
-                Text(recipe.title)
-                    .font(.system(size: 15, weight: .semibold))
-                    .foregroundStyle(Color.black)
-                    .lineLimit(1)
-
-                Text(recipe.channelName)
-                    .font(.system(size: 13, weight: .medium))
-                    .foregroundStyle(Color.gray)
-                    .lineLimit(1)
-
-                Text("조회수 \(recipe.viewCountText)")
-                    .font(.system(size: 12))
-                    .foregroundStyle(Color.gray)
-            }
-
-            Spacer()
-
-            Image(systemName: recipe.isBookmarked ? "bookmark.fill" : "bookmark")
-                .foregroundStyle(recipe.isBookmarked ? Color.black : Color.gray)
-        }
-        .padding(.vertical, 8)
-    }
-}
 
 #Preview {
     PreviewWrapper()
