+++ mohaemukzip/mohaemukzipApp.swift
@@ -11,7 +11,8 @@ import SwiftUI
 struct mohaemukzipApp: App {
     var body: some Scene {
         WindowGroup {
-            RootView()
+//            RootView()
+            MainTabView()
         }
     }
 }
